feat: 업로드 이미지 삭제 시 자동 인풋 오픈 및 useImgUpload 훅 생성
